Move createMetricsInterceptorProvider
