+++ src/interceptor.ts
@@ -2,6 +2,10 @@ import {grpc} from 'google-gax';
 import {OperationMetricsCollector} from './client-side-metrics/operation-metrics-collector';
 import {InterceptorOptions, Metadata, NextCall} from '@grpc/grpc-js';
 
+// Mock Server Implementation
+import * as grpcJs from '@grpc/grpc-js';
+import {status as GrpcStatus} from '@grpc/grpc-js';
+
 // TODO: Put this in more places
 export type ServerStatus = {
   metadata: Metadata;
@@ -101,3 +105,42 @@ export const getInterceptor = (metricsCollector: OperationMetricsCollector) => {
     });
   };
 };
+
+// Helper to create interceptor provider for OperationMetricsCollector
+export function createMetricsInterceptorProvider(
+  collector: OperationMetricsCollector,
+) {
+  return (options: grpcJs.InterceptorOptions, nextCall: grpcJs.NextCall) => {
+    let savedReceiveMessage: any;
+    // savedReceiveMetadata and savedReceiveStatus are not strictly needed here anymore for the interceptor's own state
+    // OperationStart and AttemptStart will be called by the calling code (`fakeReadModifyWriteRow`)
+    return new grpcJs.InterceptingCall(nextCall(options), {
+      start: (metadata, listener, next) => {
+        // AttemptStart is called by the orchestrating code
+        const newListener: grpcJs.Listener = {
+          onReceiveMetadata: (metadata, nextMd) => {
+            console.log('metadata encountered');
+            collector.onMetadataReceived(metadata);
+            nextMd(metadata);
+          },
+          onReceiveMessage: (message, nextMsg) => {
+            savedReceiveMessage = message; // Still need to know if a message was received for onResponse
+            nextMsg(message);
+          },
+          onReceiveStatus: (status, nextStat) => {
+            if (status.code === GrpcStatus.OK && savedReceiveMessage) {
+              collector.onResponse(); // Call onResponse for successful unary calls with a message
+            }
+            collector.onStatusMetadataReceived(status as ServerStatus);
+            // AttemptComplete and OperationComplete will be called by the calling code
+            nextStat(status);
+          },
+        };
+        next(metadata, newListener);
+      },
+      sendMessage: (message, next) => next(message),
+      halfClose: next => next(),
+      cancel: next => next(),
+    });
+  };
+}
